+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/Handlers.java
@@ -17,15 +17,22 @@
 package io.fabric8.kubernetes.client;
 
 import io.fabric8.kubernetes.api.builder.VisitableBuilder;
+import io.fabric8.kubernetes.api.model.APIResourceList;
 import io.fabric8.kubernetes.api.model.GenericKubernetesResource;
+import io.fabric8.kubernetes.api.model.GenericKubernetesResourceList;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.KubernetesResourceList;
 import io.fabric8.kubernetes.client.dsl.NamespacedInOutCreateable;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.base.HasMetadataOperation;
+import io.fabric8.kubernetes.client.dsl.base.ResourceDefinitionContext;
+import io.fabric8.kubernetes.client.utils.ApiVersionUtil;
 import io.fabric8.kubernetes.client.utils.KubernetesResourceUtil;
+import io.fabric8.kubernetes.client.utils.Serialization;
 import okhttp3.OkHttpClient;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.BiFunction;
@@ -33,6 +40,7 @@ import java.util.function.BiFunction;
 public final class Handlers {
 
   private static final Map<Class<?>, ResourceHandler<?, ?>> RESOURCE_HANDLER_MAP = new ConcurrentHashMap<>();
+  private static final Map<List<String>, ResourceDefinitionContext> GENERIC_RESOURCE_HANDLER_MAP = new ConcurrentHashMap<>();
   
   private Handlers() {
     //Utility
@@ -47,10 +55,73 @@ public final class Handlers {
   public static <T extends HasMetadata> void unregister(Class<T> type) {
     RESOURCE_HANDLER_MAP.remove(type);
   }
+  
+  /**
+   * Returns a {@link ResourceHandler} for the given item.  The client is optional, if it is supplied, then the server can be 
+   * consulted for resource metadata if a generic item is passed in.  The client is not needed at all for lookups related to 
+   * obtaining a builder - as there is no builder associated with generic resources. 
+   */
+  public static <T extends HasMetadata, V extends VisitableBuilder<T, V>> ResourceHandler<T, V> get(T meta, BaseClient client) {
+    Class<T> type = (Class<T>)meta.getClass();
+    
+    if (type.equals(GenericKubernetesResource.class)) {
+      ResourceDefinitionContext rdc = getResourceDefinitionContext((GenericKubernetesResource)meta, client);
+      
+      if (rdc != null) {
+        return new ResourceHandlerImpl(GenericKubernetesResource.class, GenericKubernetesResourceList.class, rdc);
+      }
+    }
+    
+    return get(type);
+  }
+  
+  public static ResourceDefinitionContext getResourceDefinitionContext(String apiVersion, String kind, BaseClient client) {
+    GenericKubernetesResource resource = new GenericKubernetesResource();
+    resource.setKind(kind);
+    resource.setApiVersion(apiVersion);
+    return getResourceDefinitionContext(resource, client);
+  }
 
-  public static <T extends HasMetadata, V extends VisitableBuilder<T, V>> ResourceHandler<T, V> get(T meta) {
-    // could try to return something here in the generic case, but there is no generic builder
-    return get((Class<T>)meta.getClass());
+  public static <T extends HasMetadata> ResourceDefinitionContext getResourceDefinitionContext(GenericKubernetesResource meta, BaseClient client) {
+    // check if it's built-in
+    Object value = Serialization.unmarshal(Serialization.asJson(meta));
+    Class<T> parsedType = (Class<T>) value.getClass();
+    
+    ResourceDefinitionContext rdc = null;
+    if (!parsedType.equals(GenericKubernetesResource.class)) {
+      rdc = ResourceDefinitionContext.fromResourceType(parsedType);
+    } else if (client != null) {
+      // if a client has been supplied, we can try to look this up from the server
+      String kind = meta.getKind();
+      String apiVersion = meta.getApiVersion();
+      if (kind == null || apiVersion == null) {
+        return null;
+      }
+      String api = ApiVersionUtil.trimGroupOrNull(apiVersion);
+      if (api == null) {
+        return null;
+      }
+      String version = ApiVersionUtil.trimVersion(apiVersion);
+      // assume that resource metadata won't change for the lifetime of the client
+      rdc = GENERIC_RESOURCE_HANDLER_MAP.computeIfAbsent(Arrays.asList(kind, apiVersion), k -> {
+        APIResourceList resourceList = client.getApiResources(apiVersion);
+        if (resourceList == null) {
+          return null;
+        }
+        return resourceList.getResources()
+            .stream()
+            .filter(r -> kind.equals(r.getKind()))
+            .findFirst()
+            .map(resource -> new ResourceDefinitionContext.Builder().withGroup(api)
+                .withKind(kind)
+                .withNamespaced(Boolean.TRUE.equals(resource.getNamespaced()))
+                .withPlural(resource.getName())
+                .withVersion(version)
+                .build())
+            .orElse(null);
+      });        
+    }
+    return rdc;
   }
 
   public static <T extends HasMetadata, V extends VisitableBuilder<T, V>> ResourceHandler<T, V> get(Class<T> type) {
@@ -62,6 +133,9 @@ public final class Handlers {
   
   public static <T extends HasMetadata, L extends KubernetesResourceList<T>, R extends Resource<T>> HasMetadataOperation<T, L, R> getOperation(Class<T> type, Class<L> listType, OkHttpClient client, Config config) {
     ResourceHandler<T, ?> resourceHandler = get(type);
+    if (resourceHandler == null) {
+      throw new IllegalStateException();
+    }
     return (HasMetadataOperation<T, L, R>) resourceHandler.operation(client, config, listType);
   }